:tractor: relocate deprecated code to deprecated folder and will start remove later
