+++ tests/deprecated/test_repo.py
@@ -2,7 +2,7 @@ import fs.path
 from mock import patch
 from nose.tools import eq_, raises
 from moban.exceptions import NoGitCommand
-from moban.definitions import GitRequire
+from moban.deprecated import GitRequire
 from moban.deprecated.repo import (
     git_clone,
     get_repo_name,
