+++ tests/mobanfile/test_mobanfile.py
@@ -1,8 +1,8 @@
 import fs.path
 from mock import patch
 from nose.tools import eq_
-from moban.definitions import GitRequire, TemplateTarget
-
+from moban.definitions import TemplateTarget
+from moban.deprecated import GitRequire
 
 class TestFinder:
     def setUp(self):
