+++ tests/test_definitions.py
@@ -1,5 +1,6 @@
 from nose.tools import eq_
-from moban.definitions import GitRequire, TemplateTarget
+from moban.definitions import TemplateTarget
+from moban.deprecated import GitRequire
 
 
 def test_git_require_repr():
